completed security module

+++ src/main/java/com/blog/controller/CommentController.java
@@ -1,8 +1,11 @@
 package com.blog.controller;
 
 import com.blog.model.Comment;
+import com.blog.repository.UserRepository;
 import com.blog.service.CommentServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,9 +20,11 @@ public class CommentController {
 
     @Autowired
     private final CommentServiceImpl commentService;
+    private final UserRepository userRepository;
 
-    public CommentController(CommentServiceImpl commentService) {
+    public CommentController(CommentServiceImpl commentService, UserRepository userRepository) {
         this.commentService = commentService;
+        this.userRepository = userRepository;
     }
 
     @PostMapping("/savecomment/{id}")
@@ -28,15 +33,20 @@ public class CommentController {
         comment.setPostId(postId);
         comment.setCreatedAt(new Date());
         comment.setUpdatedAt(new Date());
-        comment.setUserId(1L);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        comment.setUserId(userRepository.getUserByEmail(authentication.getName()).getId());
         commentService.saveComment(comment);
         return "redirect:/post/" + postId;
     }
 
     @GetMapping("/deletecomment/{commentId}")
     public String deleteComment(@PathVariable(value = "commentId") Long commentId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Comment comment = commentService.getCommentById(commentId);
         long postId = commentService.getCommentById(commentId).getPostId();
-        commentService.deleteComment(commentId);
+        if(userRepository.getUserByEmail(authentication.getName()).getId() == comment.getUserId()){
+            commentService.deleteComment(commentId);
+        }
         return "redirect:/post/" + postId;
     }
 
@@ -51,13 +61,16 @@ public class CommentController {
     @PostMapping("/updatecomment/{commentId}")
     public String updateComments(@PathVariable(value = "commentId") Long commentId, @ModelAttribute("commentBind")
             Comment userComment) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         Comment comment = commentService.getCommentById(commentId);
         userComment.setUserId(comment.getUserId());
         userComment.setCreatedAt(comment.getCreatedAt());
         userComment.setPostId(comment.getPostId());
         userComment.setId(commentId);
         userComment.setUpdatedAt(new Date());
-        commentService.saveComment(userComment);
+        if(userRepository.getUserByEmail(authentication.getName()).getId() == comment.getUserId()){
+            commentService.saveComment(userComment);
+        }
         return "redirect:/post/" + userComment.getPostId();
     }
 }
